Acciones de premium end

+++ src/controllers/products.controller.js
@@ -10,11 +10,15 @@ const handleCreateProduct = async (req, res, next) => {
     try {
         const socketServer = req.app.get("socketServer");
         const thumbnail = req.file?.originalname ?? "default.png";
+        const { role, email } = req.session.user;
+        const owner = role === "ADMIN" ? "ADMIN" : email;
         const productData = {
             ...req.body,
+            owner,
             thumbnail
         }
         const data = await productService.createProduct(productData);
+        req.logger.info(`Nuevo producto creado por ${email} role:${role}`)
         sendResponse(res, 201, true, "Producto creado exitosamente.", data);
         socketServer.emit("getProducts", await productService.getPaginatedProducts(3, 1, '{}', 1));/**mejorar*/
     } catch (error) {
@@ -26,20 +30,22 @@ const handleCreateProduct = async (req, res, next) => {
 
 }
 
-const handleDeleteProductRequest = async (req, res) => {
+const handleDeleteProductRequest = async (req, res, next) => {
     try {
         const socketServer = req.app.get("socketServer")
         const { productId } = req.params;
-        const data = await productService.deleteProduct(productId);
-        //res.status(200).send(data);
+        const { email, role } = req.session.user;
+        const product = await productService.getProductById(productId);
+        if (role != "ADMIN" && product.owner != email) {
+            req.logger.warning(`${email} intenta eliminar el producto ${productId}`);
+            throw new CustomError(errorsDictionary.AUTHORIZATION_ERROR, { message: "No cuentas con lo permisos necesario para eliminar este producto" })
+        }
+        await productService.deleteProduct(productId);
+        req.logger.info(`Usuario ${email} ${role} elimino el producto ${productId} `)
         sendResponse(res, 200, true, "Producto eliminado");
         socketServer.emit("getProducts", await productService.getPaginatedProducts(3, 1, '{}', 1));/**mejorar*/
-    } catch ({ message }) {
-        console.error('Error al crear producto', message);
-        const errorData = {
-            error: message,
-        };
-        sendResponse(res, 500, false, 'Error en el servidor', errorData);
+    } catch (error) {
+        next(error);
     }
 }
